rename get positions file
