+++ src/components/Copilot/utils/calculatePositions.ts
